feat(expense-detail): restructure expense detail paths, add translation and constants, and implement update expense screen
