+++ app/(tabs)/(home)/expenses/expense-detail/ExpenseDetail.translate.ts
